Add new optimization and option to enable it
The ray tracer can now pre compute where the ray could go and then pick
random directions from that list, to prevent ray misses. This is
currently less efficient than the normal uniform distribution algorithm
and can be enabled with the help of a CLI flag. The goal is to improve
that idea over time.