Return black when recursion level exceeds limit
We don't really know what color to return here because the actual color
could have been anything. Instead of returning the background/sky color,
return black which means that no intensity will be added to the original
ray.

Also, move the check up to do it before generating a random ray and
calculating `cos(theta)` to save a few milliseconds of render time.